requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
icalendar==5.0.11

